+++ src/main/java/com/elmakers/mine/bukkit/plugins/spells/builtin/FireballSpell.java
@@ -1,13 +1,14 @@
 package com.elmakers.mine.bukkit.plugins.spells.builtin;
 
-import net.minecraft.server.EntityFireball;
 import net.minecraft.server.EntityLiving;
+import net.minecraft.server.WorldServer;
 
 import org.bukkit.Location;
 import org.bukkit.Material;
 import org.bukkit.block.Block;
 import org.bukkit.craftbukkit.CraftWorld;
 import org.bukkit.craftbukkit.entity.CraftPlayer;
+import org.bukkit.entity.Entity;
 import org.bukkit.entity.Player;
 
 import com.elmakers.mine.bukkit.plugins.spells.Spell;
@@ -16,35 +17,60 @@ import com.elmakers.mine.bukkit.plugins.spells.utilities.Vec3D;
 
 public class FireballSpell extends Spell {
 
-	@Override
-	public boolean onCast(String[] parameters) 
+	protected int defaultSize = 1;
+
+	public FireballSpell()
 	{
-		Block target = getTargetBlock();
-		Location playerLoc = player.getLocation();
+		addVariant("tnt", Material.TNT, getCategory(), "Create a rather large explosion", "4");
+		addVariant("kaboom", Material.GRILLED_PORK, getCategory(), "Create a big explosion", "8");
+		addVariant("nuke", Material.BED, getCategory(), "Create a rediculous explisoin", "20");
+	}
+	
+	public boolean createExplosionAt(Location target, float size)
+	{
 		if (target == null) 
 		{
 			castMessage(player, "No target");
 			return false;
 		}
 		
-		double dx = target.getX() - playerLoc.getX();
-		double height = 1;
-        double dy = (target.getY() + (double)(height / 2.0F)) - (playerLoc.getY() + (double)(height / 2.0F));
-        double dz = target.getZ() - playerLoc.getZ();
-		
 		castMessage(player, "FOOM!");
 		CraftPlayer craftPlayer = (CraftPlayer)player;
 		EntityLiving playerEntity = craftPlayer.getHandle();
-		EntityFireball fireball = new EntityFireball(((CraftWorld)player.getWorld()).getHandle(), playerEntity, dx, dy, dz);
+        WorldServer world = ((CraftWorld)player.getWorld()).getHandle();
 		
-		double d8 = 4D;
-        Vec3D vec3d = getLocation(player, 1.0F);
-        fireball.locX = playerLoc.getX() + vec3d.xCoord * d8;
-        fireball.locY = playerLoc.getY() + (double)(height / 2.0F) + 0.5D;
-        fireball.locZ = playerLoc.getZ() + vec3d.zCoord * d8;
-        
-        ((CraftWorld)player.getWorld()).getHandle().a(fireball);
+        world.createExplosion(playerEntity, target.getBlockX(), target.getBlockY(), target.getBlockZ(), size, true);
 		return true;
+	}
+	
+	@Override
+	public boolean onCast(String[] parameters) 
+	{
+        float size = defaultSize;
+        if (parameters.length > 0)
+        {
+        	try
+        	{
+        		size = Float.parseFloat(parameters[0]);
+        	}
+        	catch (NumberFormatException ex)
+        	{
+        		size = defaultSize;
+        	}
+        }
+		Entity entity = getTargetEntity();
+		if (entity == null)
+		{
+			Block targetBlock = getTargetBlock();
+			if (targetBlock == null)
+			{
+				sendMessage(player, "No target");
+				return false;
+			}
+			return createExplosionAt(targetBlock.getLocation(), size);
+		}
+		
+		return createExplosionAt(entity.getLocation(), size);
 	}
 
 	@Override